Fix FireFox errors for initial support.
Added:
- "browser_specific_settings" for gecko including "data_collection_permissions"
- an entry for background scripts (same as service worker)

This allows the extension to be loaded in IceCat.

The quick-switch window opens and all the functions under that work  except
opening the "Manage spaces" window.  However, one can get to that page by going
to manage the extension and clicking preferences.  Maybe that's the system.display
issue?  This may only be used for opening the window somewhere specific; IDK.
Need to investigate.

Export works, however import seems to onlysort of work from a spaces-backup.JSON
copy paste.